Clean up code and two minor fixes
This commit mostly just cleans up a lot of little things
like spelling, comment style, spacing and general code
clarity for anyone who wants to read it in the future
(ourselves included).

Other than that, I added more highlighting styles to
optionally choose from in the index.html. It is as simple
as commenting in/out the ones you want to try (and I added
some comments and links for anyone who wants to explore it
more). I also tried a different style for this branch that
I thought went better with the style of our site, but we
can really use whatever we like. Aside from that there were
only two "real" changes:

- on line 97 of studyPageJS: `... overflow: hidden;
overflow-y: scroll;");` I added two more styles to work in
concert with the change on line 17 of styles.css. This
change allows cards to behave better when studying on a
small screen. Rather than a piece of code overflowing it
now is hidden and scrollable. The webkit piece is to remove
the actual scrollbar, which renders by default when
overflow is set to auto or scroll.

I also changed line 10 of studyPage.html by uncommenting it.
It appeared to have been commented out by accident while
removing the correct/incorrect buttons, and it doesn't
appear to change the functionality of the page and so it
seemed like it should be ok to uncomment it.